Move loadbalancer used data package position
