+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/LoadData.java
@@ -21,10 +21,10 @@ package org.apache.pulsar.broker.loadbalance;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
-import org.apache.pulsar.broker.BrokerData;
-import org.apache.pulsar.broker.BundleData;
 import org.apache.pulsar.broker.namespace.NamespaceService;
 import org.apache.pulsar.common.naming.NamespaceBundle;
+import org.apache.pulsar.policies.data.loadbalancer.BrokerData;
+import org.apache.pulsar.policies.data.loadbalancer.BundleData;
 
 /**
  * This class represents all data that could be relevant when making a load management decision.
